Download Soilgrids Data
I created the class Soilgirds, which currently contains the local paths to the
soilgrids files and some functions to organise the download if necessary.

The function download_file() in the Helper class performs the download.

The download works at the moment only for the 5000m and 1000m resolution.

The data is stored in the directory: soilgrids/{resolution}.

The resolution must be specified in the domain file by the variable sg_res.
sg_res = 'coarse'  -> 5000m
sg_res = 'medium'  -> 1000m
sg_res = 'fine'    ->  250m; not working