Mask soilgrids data on gridbox
I calculate for each model gridbox which pixels of the soilgrids data are
inside. Then I create a boolean mask that removes all values outside the
gridbox. a pixel is cosidered as inside if it is 50% insede the box.
This function should also work for triangular grid boxes, but it has not
been tested yet.

Problems may occur if the resolution of soilgrids is coarser than the
model grid, for which there is no exception yet.